Fixed all norm errors
