+++ includes/commands.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   commands.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ffilipe- <ffilipe-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/28 15:35:52 by paulorod          #+#    #+#             */
-/*   Updated: 2023/09/25 16:55:58 by ffilipe-         ###   ########.fr       */
+/*   Updated: 2023/09/26 13:29:31 by paulorod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,5 +39,6 @@ int		get_cmd_size(char **tokens);
 t_cmd	*set_next_cmd(t_cmd **command, int *j, int i, char **tokens);
 t_cmd	*create_token_cmd(char *token);
 void	*unclosed_quotes_error(t_var_ext *var_ext);
+char	set_new_command(char *command, int *j, int i);
 
 #endif
