+++ srcs/redirections/handle_redirections.c
@@ -6,7 +6,7 @@
 /*   By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/30 13:55:29 by ffilipe-          #+#    #+#             */
-/*   Updated: 2023/09/14 12:44:59 by paulorod         ###   ########.fr       */
+/*   Updated: 2023/09/26 13:31:24 by paulorod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,4 +57,17 @@ void	handle_redir_out_append(t_shell *shell)
 		exit(127);
 	}
 	swap_fd(&shell->cmd->std.out, fd);
+}
+
+/*Get env variable*/
+char	*get_var(char *str)
+{
+	int		i;
+	char	*var;
+
+	i = 1;
+	while (var_char_valid(str[i]))
+		i++;
+	var = ft_substr(str, 0, i);
+	return (var);
 }
